fix: Move from src to root, to have proper autocomplition
