+++ packages/common/index.ts
@@ -3,3 +3,4 @@ export * from './utils/logger';
 export * from './utils/platforms';
 export * from './utils/agruments';
 export * from './utils/sleep';
+export * from './utils/config';
