+++ packages/tosu/src/entities/AllTimesData/index.ts
@@ -1,5 +1,3 @@
-import { wLogger } from '@tosu/common';
-
 import { DataRepo } from '@/entities/DataRepoList';
 
 import { AbstractEntity } from '../AbstractEntity';
