+++ packages/tosu/src/entities/BeatmapPpData/index.ts
@@ -1,6 +1,5 @@
 import { Beatmap, Calculator } from '@kotrikd/rosu-pp';
-import { wLogger } from '@tosu/common';
-import { config } from '@tosu/common/dist/config';
+import { config, wLogger } from '@tosu/common';
 import { Beatmap as ParsedBeatmap } from 'osu-classes';
 import { BeatmapDecoder } from 'osu-parsers';
 import path from 'path';
