+++ packages/updater/tsconfig.json
@@ -1,12 +1,14 @@
 {
     "compilerOptions": {
-        "lib": ["ES2020"],
+        "lib": [
+            "ES2020"
+        ],
         "module": "commonjs",
         "moduleResolution": "Node",
         "allowJs": true,
         "esModuleInterop": true,
         "outDir": "dist",
-        "rootDir": "src",
+        "rootDir": "./",
         "sourceMap": false,
         "declaration": false,
         "strict": true,
@@ -15,6 +17,11 @@
         "strictPropertyInitialization": false,
         "baseUrl": ".",
     },
-    "exclude": ["node_modules"],
-    "include": ["src"]
+    "exclude": [
+        "node_modules",
+        "dist",
+    ],
+    "include": [
+        "**/*"
+    ],
 }
